+++ scripts/docs_version.py
@@ -1,0 +1,4 @@
+import mkdocs_gen_files
+
+with mkdocs_gen_files.open("foo.md", "w") as f:
+    print("Hello, world!", file=f)